+++ src/data/spells.js
@@ -1,9 +1,9 @@
 export const FEAT_SPELLCASTING = [
-  { name: "Kitsune Spell Familiarity", knownSpells: ["Daze", "Forbidding Ward", "Ghost Sound"], preparePBSpells: true, },
-  { name: "Kitsune Spell Expertise", knownSpells: ["Confusion", "Death Ward", "Illusory Scene"], preparePBSpells: true, },
-  { name: "Kitsune Spell Mysteries", knownSpells: ["Bane", "Illusory Object", "Sanctuary"], preparePBSpells: true, },
-  { name: "Nagaji Spell Familiarity", knownSpells: ["Daze", "Detect Magic", "Mage Hand"], preparePBSpells: true, },
-  { name: "Nagaji Spell Expertise", knownSpells: ["Blink", "Control Water", "Subconscious Suggestion"], preparePBSpells: true, },
-  { name: "Nagaji Spell Mysteries", knownSpells: ["Charm", "Fleet Step", "Heal"], preparePBSpells: true, },
+  { name: "Kitsune Spell Familiarity", showSlotless: false, knownSpells: ["Daze", "Forbidding Ward", "Ghost Sound"], preparePBSpells: true, },
+  { name: "Kitsune Spell Expertise", showSlotless: false, knownSpells: ["Confusion", "Death Ward", "Illusory Scene"], preparePBSpells: true, },
+  { name: "Kitsune Spell Mysteries", showSlotless: false, knownSpells: ["Bane", "Illusory Object", "Sanctuary"], preparePBSpells: true, },
+  { name: "Nagaji Spell Familiarity", showSlotless: false, knownSpells: ["Daze", "Detect Magic", "Mage Hand"], preparePBSpells: true, },
+  { name: "Nagaji Spell Expertise", showSlotless: false, knownSpells: ["Blink", "Control Water", "Subconscious Suggestion"], preparePBSpells: true, },
+  { name: "Nagaji Spell Mysteries", showSlotless: false, knownSpells: ["Charm", "Fleet Step", "Heal"], preparePBSpells: true, },
 ];